decrease the texture size for ssao to 16f

+++ t_ssao.go
@@ -62,7 +62,7 @@ func NewSSAO() *SsaoFBO {
 	}
 	gl.GenTextures(1, &ssao.noiseTexture)
 	gl.BindTexture(gl.TEXTURE_2D, ssao.noiseTexture)
-	gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGB32F, 4, 4, 0, gl.RGB, gl.FLOAT, gl.Ptr(&noise[0]))
+	gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGB16F, 4, 4, 0, gl.RGB, gl.FLOAT, gl.Ptr(&noise[0]))
 	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST)
 	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST)
 	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.REPEAT)
